git-svn-id: http://192.168.20.11/sequoiadb/trunk@13521 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dps/dpsLogPage.cpp
@@ -79,7 +79,7 @@ namespace engine
    {
       stringstream ss ;
       ss << "PageNumber: " << _pageNumber
-         << ", StartPage: " << _startPage
+         << ", StartPage: " << (INT32)_startPage
          << ", BeginLSNV" << _beginLSN.version
          << ", BeginLSNO" << _beginLSN.offset
          << ", Length: " << getLength()
